Fix bootstrap sample
