+++ bootstrap/src/main/AndroidManifest.xml
@@ -15,12 +15,12 @@
 
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.xfinity.bootstrap">
+    package="com.xfinity.blueprint_bootstrap">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="com.xfinity.blueprint.bootstrap.BootstrapApplication"
+        android:name=".BootstrapApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -28,7 +28,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.xfinity.blueprint.bootstrap.MainActivity"
+            android:name=".MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
