+++ bootstrap/src/main/java/com/xfinity/blueprint_bootstrap/component/model/SimpleTextModel.kt
@@ -12,16 +12,8 @@
  *
  */
 
-package com.xfinity.blueprint.bootstrap.webservices
+package com.xfinity.blueprint_bootstrap.component.model
 
-import okhttp3.Interceptor
-import okhttp3.Response
+import com.xfinity.blueprint.model.ComponentModel
 
-class ApiKeyInterceptor(private val apiKey: String) : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        var request = chain.request()
-        val url = request.url().newBuilder().addQueryParameter("appid", apiKey).build()
-        request = request.newBuilder().url(url).build()
-        return chain.proceed(request)
-    }
-}
+data class SimpleTextModel(val text: CharSequence) : ComponentModel
