+++ bootstrap/src/main/java/com/xfinity/blueprint_bootstrap/model/api/CurrentWeather.kt
@@ -12,16 +12,7 @@
  *
  */
 
-package com.xfinity.blueprint.bootstrap.component.presenter
+package com.xfinity.blueprint_bootstrap.model.api
 
-import com.xfinity.blueprint.bootstrap.component.model.SimpleTextModel
-import com.xfinity.blueprint.bootstrap.component.view.HelloComponent
-import com.xfinity.blueprint.presenter.ComponentPresenter
-import com.xfinity.blueprint_annotations.DefaultPresenter
-
-@DefaultPresenter(viewClass = HelloComponent::class)
-class HelloComponentPresenter : ComponentPresenter<HelloComponent, SimpleTextModel> {
-    override fun present(view: HelloComponent, model: SimpleTextModel) {
-        view.setHelloText(model.text)
-    }
-}
+data class CurrentWeather(val weather: List<Weather>)
+data class Weather(val main: String, val description: String)
